requirements.txt
click==8.1.3
colorama==0.4.5
dataclasses==0.6
Flask==2.2.2
Flask-SQLAlchemy==3.0.0
greenlet==1.1.3
invoke==1.7.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2-binary==2.9.3
python-dotenv==0.21.0
pywin32-ctypes==0.2.0
rfc3986==1.5.0
SQLAlchemy==1.4.41
webencodings==0.5.1
Werkzeug==2.2.2
zipp==3.10.0